moved method to seperate class
